Modify: Agent creation logic
A new entry to the contacts table will trigger a PL/SQL to create and update agents.
For a given organization a new contact will create a new agent with rank 1,
All the existing agent ranks in the organization will be incremented 1.
Agent segment_name values will be updated dynamically by updated ranks.